lib.mdDoc has been deprecated since 24.05
`lib.mdDoc` emits this trace when called since 24.05:

```quote
trace: warning: lib.mdDoc will be removed from nixpkgs in 24.11. Option descriptions are now in Markdown by default; you can remove any remaining uses of lib.mdDoc.
```

Given that this was introduced with 22.11 and made obsolete in 24.05 with all prior versions being EoL at this point it should be safe to revert that change.
In any case the only result will be that anything ≥22.11 and <24.05 will be rendered without markup which I guess isn't the best, but given those are EoL systems it's not the worst.

Either way it gets rid of a *lot* of noise for 24.05 users in the trace, and for 24.11 and up it will remove dead code.

This reverts commit 7602e548a78932bd28a7e2f621b3d62b4124e993.

Signed-off-by: benaryorg <binary@benary.org>