flake: fix patch for nix-eval-jobs
We carry a custom patch for nix-eval-jobs that causes the tool to not
unset the `NIX_PATH` environment variable anymore. The patch does not
apply anymore starting with d0bcdae (clang-tidy: use trailing return
type, 2024-11-10) because surrounding code has changed. Other than that
the patch still applies.

The mentioned commit was released with nix-eval-jobs v2.25.0, which
nixos-unstable has recently upgraded to. Adapt the patch to unbreak the
build.

Fixes #255.