command/apply: distinguish default and explicit "--no-substitutes"
In the next commit, we're about to introduce a new per-node option that
disables substitutes by default. If set, the "--no-substitutes" command
line option shall continue to override that default, following our usual
rules of preference.

In order to override the default, we're also going to introduce a new
"--use-substitutes" option. But this will require us to distinguish the
cases where either of these flags has been set from the case where none
of them were passed:

  - "--no-substitutes" overrides the per-node option and disables
    substitutes.

  - "--use-substitutes overrides the per-node option and enables
    substitutes.

  - No flag will cause the per-node option to kick in.

Refactor the code so that we use an `Option<bool>`. This allows us to
easily discern the last case in subsequent commits.